Adicionei o get user id nos comprovantes

+++ src/main/java/dispesas/com/service/ComprovanteService.java
@@ -4,6 +4,7 @@ import dispesas.com.Repository.ComprovanteRepository;
 import dispesas.com.Repository.DespesaRepository;
 import dispesas.com.model.Comprovante;
 import dispesas.com.model.Despesa;
+import dispesas.com.security.utilSecurity.GetUserById;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
@@ -21,6 +22,7 @@ import java.util.List;
 public class ComprovanteService {
     private  final DespesaRepository despesaRepository;
     private final ComprovanteRepository comprovanteRepository;
+    private final GetUserById getUserById;
 
     @Transactional
     public void salvarComprovante(MultipartFile file, Long idDespesa){
@@ -32,7 +34,11 @@ public class ComprovanteService {
         if (file.getSize() > 5L * 1024 * 1024){
             throw new RuntimeException("Arquivo muito grande. Máximo 5MB");
         }
+        Long userId = getUserById.getUserById().getId();
         Despesa despesa = despesaRepository.findById(idDespesa).orElseThrow(() -> new RuntimeException("Nenhum despesa encontrada"));
+        if (!despesa.getUser().getId().equals(userId)){
+            throw new RuntimeException("Erro: Usurio não é o correto");
+        }
 
         try{
             Comprovante comprovante = new Comprovante();
@@ -57,6 +63,11 @@ public class ComprovanteService {
     //Listar comprovante por id
     public ResponseEntity<byte[]> visualizarComprovante(Long idComprovante){
        Comprovante comprovante  = comprovanteRepository.findById(idComprovante).orElseThrow(() -> new RuntimeException("Erro: Nenhum comprovante localizado com este id"));
+        Long userId = getUserById.getUserById().getId();
+        if (!comprovante.getDespesa().getUser().getId().equals(userId)){
+           throw new RuntimeException("Erro: Usurio não é o correto");
+       }
+
         byte[] data = comprovante.getFile();
         if (data == null || data.length == 0){
             return ResponseEntity.noContent().build();
@@ -86,9 +97,17 @@ public class ComprovanteService {
 
     @Transactional
     public void deletarComprovante(Long id){
+        Comprovante comprovante  = comprovanteRepository.findById(id).orElseThrow(() -> new RuntimeException("Erro: Nenhum comprovante localizado com este id"));
+        Long userId = getUserById.getUserById().getId();
+
+        if (!comprovante.getDespesa().getUser().getId().equals(userId)){
+            throw new RuntimeException("Erro: Usurio não é o correto");
+        }
+
         if (!comprovanteRepository.existsById(id)){
             throw new RuntimeException("Comprovante não existe");
         }
+
         comprovanteRepository.deleteById(id);
     }
